finished admin route

+++ src/controllers/adminController.ts
@@ -1,11 +1,19 @@
 
+import { log } from "console";
 import {Request, Response} from "express";
-import { addAdminService, loginAdminService, getAllAdminService, getSingleAdminService, updateAdminService, changeAccessService, blockAdminService, deleteAdminService} from "../services/adminService"
+import { addAdminService, loginAdminService, getAllAdminService, getSingleAdminService, updateAdminService, changeAccessService, blockAdminService, deleteAdminService, changePasswordService, resetPasswordEmailService, resetPasswordEmailValidateService} from "../services/adminService"
 
 export const addAdmin = async(req: Request, res: Response) => {
     try {
-        const response = await addAdminService(req.body)
-        return res.status(response.status).json(response)
+        if(req.user){
+            const response = await addAdminService(req.body)
+            return res.status(response.status).json(response)
+        }else{
+            return res.status(400).json({
+                status:400,
+                message: "Unauthorized User"
+            })
+        }
     } catch (error) {
         return res.status(500).json(error)
     }
@@ -22,8 +30,15 @@ export const loginAdmin = async(req: Request, res: Response) => {
 
 export const allAdmin = async(req: Request, res: Response) => {
     try{
-        const response = await getAllAdminService()
-        return res.status(response.status).json(response)
+        if(req.user){
+            const response = await getAllAdminService()
+            return res.status(response.status).json(response)
+        }else{
+            return res.status(401).json({
+                status:401,
+                message: "Unauthorized User"
+            })
+        }
     }catch(err){
         return res.status(500).json(err)
     }
@@ -32,8 +47,15 @@ export const allAdmin = async(req: Request, res: Response) => {
 export const singleAdmin = async(req: Request, res: Response) => {
     try {
         const {id:adminId} = req.params
-        const response = await getSingleAdminService(adminId)
-        return res.status(response.status).json(response)
+        if(req.user === adminId){
+            const response = await getSingleAdminService(adminId)
+            return res.status(response.status).json(response)
+        }else{
+            return res.status(401).json({
+                status:401,
+                message: "Unauthorized User"
+            })
+        } 
     } catch (error) {
         return res.status(500).json(error)
     }
@@ -42,8 +64,15 @@ export const singleAdmin = async(req: Request, res: Response) => {
 export const updateAdmin = async(req:Request, res: Response) => {
     try {
         const {id:adminId} = req.params
-        const response = await updateAdminService(adminId, req.body)
-        return res.status(response.status).json(response)
+        if(req.user === adminId){
+            const response = await updateAdminService(adminId, req.body)
+            return res.status(response.status).json(response)
+        }else{
+            return res.status(401).json({
+                status:401,
+                message: "Unauthorized User"
+            })
+        }
     } catch (error) {
         return res.status(500).json(error)
     }
@@ -52,8 +81,15 @@ export const updateAdmin = async(req:Request, res: Response) => {
 export const changeAccess = async(req:Request, res: Response) => {
     try {
         const {id:adminId} = req.params
-        const response = await changeAccessService(adminId, req.body)
-        return res.status(response.status).json(response)
+        if(req.user === adminId){
+            const response = await changeAccessService(adminId, req.body)
+            return res.status(response.status).json(response)
+        }else{
+            return res.status(401).json({
+                status:401,
+                message: "Unauthorized User"
+            })
+        }
     } catch (error) {
         return res.status(500).json(error)
     }
@@ -62,8 +98,15 @@ export const changeAccess = async(req:Request, res: Response) => {
 export const blockAdmin = async(req:Request, res: Response) => {
     try {
         const {id:adminId} = req.params
-        const response = await blockAdminService(adminId)
-        return res.status(response.status).json(response)
+        if(req.user === adminId){
+            const response = await blockAdminService(adminId)
+            return res.status(response.status).json(response)
+        }else{
+            return res.status(401).json({
+                status:401,
+                message: "Unauthorized User"
+            })
+        }
     } catch (error) {
         return res.status(500).json(error)
     }
@@ -72,8 +115,65 @@ export const blockAdmin = async(req:Request, res: Response) => {
 export const deleteAdmin = async(req:Request, res: Response) => {
     try {
         const {id:adminId} = req.params
-        const response = await deleteAdminService(adminId)
-        return res.status(response.status).json(response)
+        if(req.user === adminId){
+            const response = await deleteAdminService(adminId)
+            return res.status(response.status).json(response)
+        }else{
+            return res.status(401).json({
+                status:401,
+                message: "Unauthorized User"
+            })
+        }
+    } catch (error) {
+        return res.status(500).json(error)
+    }
+};
+export const changePasswordAdmin = async(req:Request, res: Response) => {
+    try {
+        const {id:adminId} = req.params
+        if(req.user === adminId){
+            const response = await changePasswordService(adminId, req.body)
+            return res.status(response.status).json(response)
+        }else{
+            return res.status(401).json({
+                status:401,
+                message: "Unauthorized User"
+            })
+        }
+    } catch (error) {
+        return res.status(500).json(error)
+    }
+};
+
+export const resetPasswordEmailAdmin = async(req:Request, res: Response) => {
+    try {
+        const {id:adminId} = req.params
+        if(req.user === adminId){
+            const response = await resetPasswordEmailService(adminId, req.body)
+            return res.status(response.status).json(response)
+        }else{
+            return res.status(401).json({
+                status:401,
+                message: "Unauthorized User"
+            })
+        }
+    } catch (error) {
+        return res.status(500).json(error)
+    }
+};
+
+export const resetPasswordEmailValidateAdmin = async(req:Request, res: Response) => {
+    try {
+        const {id:adminId} = req.params
+        if(req.user === adminId){
+            const response = await resetPasswordEmailValidateService(adminId, req.body)
+            return res.status(response.status).json(response)
+        }else{
+            return res.status(401).json({
+                status:401,
+                message: "Unauthorized User"
+            })
+        }
     } catch (error) {
         return res.status(500).json(error)
     }
